Refactor away the bstradd module
The bstradd module is made up of a few wrappers and extensions for the bstring library

However, these symbols are very sparsely used in the codebase
and are better refactored or moved closer to where they're used

At the same time, shore up error handling around these bstring API calls
