+++ bin/ad/ad_mv.c
@@ -29,7 +29,6 @@
 #include <unistd.h>
 
 #include <atalk/adouble.h>
-#include <atalk/bstradd.h>
 #include <atalk/bstrlib.h>
 #include <atalk/ftw.h>
 #include <atalk/queue.h>
