+++ etc/afpd/acls.c
@@ -39,7 +39,6 @@
 #include <atalk/acl.h>
 #include <atalk/adouble.h>
 #include <atalk/afp.h>
-#include <atalk/bstradd.h>
 #include <atalk/bstrlib.h>
 #include <atalk/cnid.h>
 #include <atalk/errchk.h>
