+++ etc/afpd/fork.c
@@ -19,7 +19,6 @@
 #include <atalk/adouble.h>
 #include <atalk/afp.h>
 #include <atalk/asp.h>
-#include <atalk/bstradd.h>
 #include <atalk/bstrlib.h>
 #include <atalk/cnid.h>
 #include <atalk/dsi.h>
