+++ etc/afpd/ofork.c
@@ -15,7 +15,6 @@
 #include <sys/stat.h> /* works around a bug */
 #include <unistd.h>
 
-#include <atalk/bstradd.h>
 #include <atalk/bstrlib.h>
 #include <atalk/ea.h>
 #include <atalk/fce_api.h>
