+++ etc/netatalk/netatalk.c
@@ -32,7 +32,6 @@
 
 #include <atalk/adouble.h>
 #include <atalk/afp.h>
-#include <atalk/bstradd.h>
 #include <atalk/bstrlib.h>
 #include <atalk/compat.h>
 #include <atalk/dsi.h>
@@ -99,7 +98,16 @@ static int set_sl_volumes(void)
     for (vol = volumes; vol; vol = vol->v_next) {
         if (vol->v_flags & AFPVOL_SPOTLIGHT) {
             bstring volnamequot = bformat("'%s'", vol->v_path);
-            bstrListPush(vollist, volnamequot);
+
+            if (vollist->qty == vollist->mlen
+                    && bstrListAlloc(vollist, vollist->qty + 1) != BSTR_OK) {
+                LOG(log_error, logtype_default,
+                    "set_sl_volumes: failed to initialize indexing for %s",
+                    bdata(volnamequot));
+            }
+
+            vollist->entry[vollist->qty] = volnamequot;
+            vollist->qty++;
         }
     }
