+++ include/atalk/util.h
@@ -139,6 +139,15 @@ extern void mod_close(void *);
 #define strequal(a,b) (strcmp((a),(b)) == 0)
 #endif
 
+#define cfrombstr(b) ((char *)((b)->data))
+
+/* strip slashes from end of a bstring */
+#define BSTRING_STRIP_SLASH(a)                      \
+    do {                                            \
+        while (bchar((a), blength(a) - 1) == '/')   \
+            bdelete((a), blength(a) - 1, 1);        \
+    } while (0);
+
 /******************************************************************
  * locking.c
  ******************************************************************/
