+++ libatalk/adouble/ad_conv.c
@@ -30,7 +30,6 @@
 #include <sys/param.h>
 
 #include <atalk/adouble.h>
-#include <atalk/bstradd.h>
 #include <atalk/bstrlib.h>
 #include <atalk/compat.h>
 #include <atalk/dsi.h>
