+++ libatalk/adouble/ad_flush.c
@@ -33,7 +33,6 @@
 #include <string.h>
 
 #include <atalk/adouble.h>
-#include <atalk/bstradd.h>
 #include <atalk/bstrlib.h>
 #include <atalk/ea.h>
 #include <atalk/errchk.h>
