+++ libatalk/adouble/ad_open.c
@@ -43,7 +43,6 @@
 #include <sys/param.h>
 
 #include <atalk/adouble.h>
-#include <atalk/bstradd.h>
 #include <atalk/bstrlib.h>
 #include <atalk/compat.h>
 #include <atalk/ea.h>
