+++ libatalk/adouble/ad_write.c
@@ -13,7 +13,6 @@
 #include <sys/param.h>
 
 #include <atalk/adouble.h>
-#include <atalk/bstradd.h>
 #include <atalk/bstrlib.h>
 #include <atalk/ea.h>
 #include <atalk/errchk.h>
