+++ libatalk/bstring/meson.build
@@ -1,8 +1,6 @@
-bstring_sources = ['bstradd.c', 'bstrlib.c']
-
 libbstring = static_library(
     'bstring',
-    bstring_sources,
+    ['bstrlib.c'],
     include_directories: root_includes,
     install: false,
 )
