+++ libatalk/util/cnid.c
@@ -46,7 +46,6 @@
 #include <sys/types.h>
 #include <unistd.h>
 
-#include <atalk/bstradd.h>
 #include <atalk/bstrlib.h>
 #include <atalk/cnid.h>
 #include <atalk/errchk.h>
