+++ libatalk/vfs/vfs.c
@@ -32,7 +32,6 @@
 #include <atalk/acl.h>
 #include <atalk/adouble.h>
 #include <atalk/afp.h>
-#include <atalk/bstradd.h>
 #include <atalk/bstrlib.h>
 #include <atalk/compat.h>
 #include <atalk/directory.h>
